Fix incorrect rF2Rules marshalling code and adjust FO detection to keep Rolling phase till the start.

+++ CrewChiefV4/RF2/RF2Data.cs
@@ -627,7 +627,7 @@ namespace CrewChiefV4.rFactor2
         {
             // input only
             public rF2TrackRulesCommand mCommand;        // recommended action
-            public long mID;                             // slot ID if applicable
+            public int mID;                              // slot ID if applicable
             public double mET;                           // elapsed time that event occurred, if applicable
         }
